fix(gui,sdl3)!: give the GUI/SDL native ABI declared widths (A2-0b)
The GUI and SDL3 native layers were written when every Zan integer was i64, so
their C signatures took `i64`/`zan_i64` for everything. Now that `int` is i32
all 176 of those declarations disagree with their C definitions, and the two
kinds of disagreement fail differently:

- coordinates, colors, ids and flags: the caller writes 32 bits of an argument
  register the callee reads as 64, so the upper half is whatever was there
- SDL_Window/SDL_Renderer/SDL_Texture and the SDL_GPU device/texture are
  pointers returned through `zan_i64`; Zan was catching them in `int`, which
  truncates every handle above 4 GiB and hands the runtime back a bad pointer

Each parameter now says what it is: values are `i32`/`int`, pointer handles are
`intptr_t`/`nint` (SdlWindow, SdlRenderer, SdlTexture, SdlGpu, SdlGpuTexture
carry `nint`), and the two real 64-bit quantities (`zan_gui_get_tick_ms`,
`zan_sdl_ticks`/`zan_sdl_event_timestamp`) stay `i64`/`long`. Also fixed:
zan_gui_poll_event took six arguments the C side never declared, sleep_ms is
void, destroy_surface returns int, and SDL_RestoreWindow takes a window pointer.

The audit that found these (extern declaration vs. C definition, by DllImport
EntryPoint) now reports 0 mismatches for the 245 externs we define ourselves;
system-library externs are the next batch. ctest: 695/695.

+++ src/runtime/gui_runtime_font.c
@@ -322,8 +322,8 @@ static i64 ft_measure_text(const char *text, int font_size) {
 }
 #endif
 
-EXPORT void zan_gui_draw_text(i64 surface_id, i64 x, i64 y,
-                              const char *text, i64 color, i64 font_size) {
+EXPORT void zan_gui_draw_text(
+    i32 surface_id, i32 x, i32 y, const char *text, i32 color, i32 font_size) {
 #ifdef ZAN_GUI_FREETYPE
     if (ft_prepare((int)font_size)) {
         ft_draw_text(surface_id, x, y, text, color, (int)font_size);
@@ -333,14 +333,14 @@ EXPORT void zan_gui_draw_text(i64 surface_id, i64 x, i64 y,
     bitmap_draw_text(surface_id, x, y, text, color, font_size);
 }
 
-EXPORT i64 zan_gui_measure_text(const char *text, i64 font_size) {
+EXPORT i32 zan_gui_measure_text(const char *text, i32 font_size) {
 #ifdef ZAN_GUI_FREETYPE
     if (ft_prepare((int)font_size)) return ft_measure_text(text, (int)font_size);
 #endif
     return bitmap_measure_text(text, font_size);
 }
 
-EXPORT i64 zan_gui_font_height(i64 font_size) {
+EXPORT i32 zan_gui_font_height(i32 font_size) {
 #ifdef ZAN_GUI_FREETYPE
     if (ft_prepare((int)font_size))
         return (i64)(g_ft_face->size->metrics.height >> 6);
@@ -356,7 +356,7 @@ EXPORT i64 zan_gui_font_height(i64 font_size) {
 #if defined(__linux__) && !defined(ZAN_GUI_SDL)
 /* ---- window management (EWMH / Xlib) ---- */
 
-EXPORT i64 zan_gui_minimize(iptr hwnd_val) {
+EXPORT i32 zan_gui_minimize(iptr hwnd_val) {
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_x11_window;
     if (g_display && xid) {
         XIconifyWindow(g_display, xid, DefaultScreen(g_display));
@@ -365,14 +365,14 @@ EXPORT i64 zan_gui_minimize(iptr hwnd_val) {
     return 0;
 }
 
-EXPORT i64 zan_gui_toggle_maximize(iptr hwnd_val) {
+EXPORT i32 zan_gui_toggle_maximize(iptr hwnd_val) {
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_x11_window;
     x11_wm_state(xid, x11_atom("_NET_WM_STATE_MAXIMIZED_VERT"),
                  x11_atom("_NET_WM_STATE_MAXIMIZED_HORZ"), 2 /* toggle */);
     return 0;
 }
 
-EXPORT i64 zan_gui_is_maximized(iptr hwnd_val) {
+EXPORT i32 zan_gui_is_maximized(iptr hwnd_val) {
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_x11_window;
     if (!g_display || !xid) return 0;
     Atom vert = x11_atom("_NET_WM_STATE_MAXIMIZED_VERT");
@@ -397,7 +397,7 @@ EXPORT i64 zan_gui_is_maximized(iptr hwnd_val) {
 
 /* 1 while the window can be seen (not iconified/unmapped); ambient
  * animations pause while this reports 0. */
-EXPORT i64 zan_gui_window_visible(iptr hwnd_val) {
+EXPORT i32 zan_gui_window_visible(iptr hwnd_val) {
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_x11_window;
     if (!g_display || !xid) return 0;
     XWindowAttributes attrs;
@@ -411,7 +411,7 @@ EXPORT i64 zan_gui_window_visible(iptr hwnd_val) {
 /* 1 while the window holds the input focus; ambient animations idle down to a
  * slow heartbeat while this reports 0, so a background window costs almost
  * nothing. */
-EXPORT i64 zan_gui_window_focused(iptr hwnd_val) {
+EXPORT i32 zan_gui_window_focused(iptr hwnd_val) {
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_x11_window;
     if (!g_display || !xid) return 0;
     Window focused = 0;
@@ -420,22 +420,22 @@ EXPORT i64 zan_gui_window_focused(iptr hwnd_val) {
     return focused == xid ? 1 : 0;
 }
 
-EXPORT i64 zan_gui_set_topmost(iptr hwnd_val, i64 on) {
+EXPORT i32 zan_gui_set_topmost(iptr hwnd_val, i32 on) {
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_x11_window;
     x11_wm_state(xid, x11_atom("_NET_WM_STATE_ABOVE"), 0, on ? 1 : 0);
     return 0;
 }
 
 /* ---- client-side title-bar metrics (borderless window) ---- */
-EXPORT i64 zan_gui_titlebar_height(void) { return g_titlebar_h_l; }
-EXPORT i64 zan_gui_caption_button_width(void) { return g_btn_w_l; }
-EXPORT i64 zan_gui_set_caption_buttons(iptr hwnd_val, i64 count) {
+EXPORT i32 zan_gui_titlebar_height(void) { return g_titlebar_h_l; }
+EXPORT i32 zan_gui_caption_button_width(void) { return g_btn_w_l; }
+EXPORT i32 zan_gui_set_caption_buttons(iptr hwnd_val, i32 count) {
     (void)hwnd_val;
     if (count >= 0 && count <= 8) { g_caption_btn_count_l = (int)count; }
     return 0;
 }
 
-EXPORT i64 zan_gui_close_window(iptr hwnd_val) {
+EXPORT i32 zan_gui_close_window(iptr hwnd_val) {
     if (!g_display) return 0;
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_primary_win;
     zan_lwin_t *w = lwin_find(xid);
@@ -480,7 +480,7 @@ EXPORT i64 zan_gui_close_window(iptr hwnd_val) {
 /* On Linux close_window already destroys the X window synchronously, so the
  * owner-driven destroy is a no-op kept for FFI symbol parity across
  * backends. */
-EXPORT i64 zan_gui_destroy_window(iptr hwnd_val) { (void)hwnd_val; return 0; }
+EXPORT i32 zan_gui_destroy_window(iptr hwnd_val) { (void)hwnd_val; return 0; }
 
 /* Native glass on Linux: ask a compositing WM (KWin, or picom via rules) to
  * blur whatever is behind the window by setting the de-facto standard
@@ -489,7 +489,7 @@ EXPORT i64 zan_gui_destroy_window(iptr hwnd_val) { (void)hwnd_val; return 0; }
  * which requires a 32-bit ARGB visual plus a running compositor; without those
  * the hint is a harmless no-op. tint is unused (the compositor owns the tint).
  * This is the Linux side of the same Gui.Native.Window.EnableGlass API. */
-EXPORT i64 zan_gui_enable_glass(iptr hwnd_val, i64 tint_argb) {
+EXPORT i32 zan_gui_enable_glass(iptr hwnd_val, i32 tint_argb) {
     (void)tint_argb;
     if (!g_display) return 1;
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_primary_win;
@@ -502,7 +502,7 @@ EXPORT i64 zan_gui_enable_glass(iptr hwnd_val, i64 tint_argb) {
     return 0;
 }
 
-EXPORT i64 zan_gui_disable_glass(iptr hwnd_val) {
+EXPORT i32 zan_gui_disable_glass(iptr hwnd_val) {
     if (!g_display) return 1;
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_primary_win;
     if (!xid) return 1;
@@ -514,7 +514,7 @@ EXPORT i64 zan_gui_disable_glass(iptr hwnd_val) {
 
 /* Whole-window opacity via the EWMH _NET_WM_WINDOW_OPACITY hint (honoured by
  * any compositing WM). percent is 10..100. */
-EXPORT i64 zan_gui_set_opacity(iptr hwnd_val, i64 percent) {
+EXPORT i32 zan_gui_set_opacity(iptr hwnd_val, i32 percent) {
     if (!g_display) return 1;
     Window xid = hwnd_val ? (Window)(intptr_t)hwnd_val : g_primary_win;
     if (!xid) return 1;
@@ -533,7 +533,7 @@ EXPORT i64 zan_gui_set_opacity(iptr hwnd_val, i64 percent) {
     return 0;
 }
 
-EXPORT i64 zan_gui_get_dpi_scale(void) {
+EXPORT i32 zan_gui_get_dpi_scale(void) {
     if (!g_display) return 100;
     int screen = DefaultScreen(g_display);
     int wpx = DisplayWidth(g_display, screen);
@@ -545,7 +545,7 @@ EXPORT i64 zan_gui_get_dpi_scale(void) {
     return (i64)scale;
 }
 
-EXPORT i64 zan_gui_set_clipboard(const char *utf8) {
+EXPORT i32 zan_gui_set_clipboard(const char *utf8) {
     if (!g_display || !g_x11_window) return 1;
     free(g_clip_text_linux);
     g_clip_text_linux = utf8 ? strdup(utf8) : strdup("");
@@ -621,7 +621,7 @@ static int x11_read_selection(Atom selection, Atom target, char **out) {
     }
 }
 
-EXPORT void zan_gui_set_ime_pos(i64 x, i64 y) {
+EXPORT void zan_gui_set_ime_pos(i32 x, i32 y) {
     /* X11 IME (XIM over-the-spot) not wired yet; accept + ignore. */
     (void)x; (void)y;
 }
